face-service/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Settings
pydantic-settings==2.3.4

# Face recognition
facenet-pytorch==2.5.3
torch==2.3.1
torchvision==0.18.1
Pillow==10.3.0
numpy==1.26.4

# Testing
pytest==8.2.2
httpx==0.27.0

# Optional — speeds up startup time by caching model weights
huggingface-hub==0.23.4
